AEKR-1622: Added images.tar check in offline scripts

+++ local-installation/offline-helper-scripts/start-feder8-offline.sh
@@ -15,6 +15,11 @@ else
   IS_MAC=false
 fi
 
-echo Loading docker images. This could take a while...
-docker load < images.tar
-docker run --rm -it --name feder8-installer -e CURRENT_DIRECTORY=$(pwd) -e IS_WINDOWS=false -e IS_MAC=$IS_MAC -e DOCKER_CERT_SUPPORT=$DOCKER_CERT_SUPPORT -v /var/run/docker.sock:/var/run/docker.sock ${REGISTRY}/library/install-script:${TAG} feder8 init full --offline
+if [ -f "images.tar" ]; then
+  echo Loading docker images. This could take a while...
+  docker load < images.tar
+  docker run --rm -it --name feder8-installer -e CURRENT_DIRECTORY=$(pwd) -e IS_WINDOWS=false -e IS_MAC=$IS_MAC -e DOCKER_CERT_SUPPORT=$DOCKER_CERT_SUPPORT -v /var/run/docker.sock:/var/run/docker.sock ${REGISTRY}/library/install-script:${TAG} feder8 init full --offline
+else
+  echo Could not find 'images.tar' in the current directory. Unable to continue.
+  exit 1
+fi
